use size_t to reduce warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@6908 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/VTBuilder/vtImage.cpp
@@ -804,7 +804,7 @@ int vtImage::NumBitmapsOnDisk()
 
 void vtImage::GetRGB(int x, int y, RGBi &rgb, double dRes)
 {
-	int closest_bitmap = 0;
+	size_t closest_bitmap = 0;
 	double diff = 1E9;
 
 	if (dRes != 0.0)
